PsychVRHMD: Make 'AutoSetupHMD' probe sequence more robust.
Lots of try-catches to increase the chance of getting through failed
driver startups unharmed and make it to the end of the tunnel.

Especially OpenHMD needs this, because if OpenHMD tries to open a
HMD which is already in exclusive use by a VR compositor like Monado
or SteamVR then that will error out, so we need to try-catch that and
carry on.