VRInputStuffTest: Update help text, add reference space selection.
So we can optionally test reference spaces. So far all four tested
OpenXR runtimes support the optional type 3 STAGE reference space,
so maybe it is commonly supported. The STAGE reference space weirdly
fixes the problem of the "misplaced teapot" in our demos under Linux+
SteamVR OpenXR, so maybe the teapot is misplaced because i did screw
up the "room setup" or origin calibration for local "sitting" experience?

What it also does is enable VRAreaBoundaries reporting of the inscribed
play area rectangle under both OculusXR Windows, and SteamVR on both
Linux and Windows, whereas Monado does not seem to support this yet.
Anyhow, when it is supported, it works and gives reasonable results.

The current driver is now pretty much fully tested under Linux and
Windows-10 with Monado and SteamVR on Linux, OculusVR and SteamVR on
Windows, with all our demos in most meaningful settings. However, only
tested with a Oculus Rift CV-1 and Oculus touch controllers, as i don't
have access to other VR hardware for further testing yet.

Anyhow, this now establishes a good baseline for a potential first
release, if we want to, with more testing/validation and especially
Linux+Monado specific timing improvements in store.

Work time so far: 707 hours! We are losing money with this, instead of
making it :(*