AMD/Windows: Undo gfxhwtype=5 "fix"
In 07/10/17,  gfxhwtype=5 was introduced and chosen as default for AMD cards under Windows if GPUCoreId=='R600' and ditherApiVer==2. Back then, this fixed mapping glitches in the gamma table with "newer" graphics card drivers (for which ditherApiVer==2 was supposed to be an indicator). Meanwhile, this fix is not necessary anymore. In fact, now gfxhwtype=5 shows mapping glitches.
Tested with several Win10 versions, AMD drivers, video ports (DP, DVI, DP++/DVI) and AMD cards (HD-7750, R7-250X, R7-260X, RX-550, RX-6400, WX-5100) - although not all combinations of these. The bottom line: nowadays, gfxhwtype=0 is always good, gfxhwtype=5 is mostly bad.
If gfxhwtype=5 has glitches, it is always that the gamma table output is shifted by -1 within a code range starting at somewhere between 9 to 45 and ending, if at all, at 255, consistently across the three color channels though. Which code range is affected, within the mentioned limits, depends on the graphics card, the output port, and what not else. For example, for the Radeon RX550, the codes 9-254 would map to 8-253, but only for DP and DP++/DVI, whereas all is fine with DVI, at least with an older driver. With newer drivers, the mapping is also screwed up with DVI.