OpenXR: Add experimental code for Monado metrics output parsing.
The file monadometricsparser.c can be compiled to an executable that
allows live parsing of Monado OpenXR runtime metrics protobuffer log
files, generated by Monado under Linux if it is launched with proper
options on a system with timing capable Vulkan drivers.

This is an early prototype for finding a hacky stop-gap solution for
OpenXR stimulus onset timestamping under Linux + Monado.

It uses the nanopb simple parsing library for Google protobuffers as
used by Monado for metrics logging, and some header files taken from
Monado for interop.

Not much to see here yet, just a backup commit of work in progress.