PsychOpenXRCore: Make presenter thread startup less racy.
Do PsychSignalCondition() after PsychLockMutex() to have a well-defined
startup:

1. presenter thread starts up, blocks on mutex.
2. Main thread enters PsychWaitCondition, drops mutex, waits on condition.
3. presenterThread gets mutex, signals main thread condition.
4. Main thread wakes from wait, blocks on mutex now held by presenter.
5. presenterThread can do its 1st PresentCycle and xrEndFrame() before
   dropping the mutex.
6. Main thread can lock+unlock mutex and resume ops.

With the previous order, it was dependent on execution timing if the
main thread would resume doing stuff before or after the presenter
managed to do its first PresentCycle + xrEndFrame. This caused a race
which could cause OpenXR present failure at least under Monado with the
new metrics timestamping, depending if a print statement was executed
or not, delaying main thread execution by 1-2 microseconds.

The reason for failure is still a bit unclear, but before the startup
sequence was ill defined and random, now it is well defined, so that's
a win for future reasoning, and seems to fix the startup problem with
the new metrics timestamping at verbosity level > 3.