Linux Octave build scripts: Remove use of striplibsfrommexfile().
This is no longer needed since Octave 5+ on Linux, as it strips all
those libs by itself in its mex/mkoctfile function. As Octave 5+
ships in all our currently supported distros, ie. Debian 11+ and
Ubuntu 20.04-LTS and later, no need for this anymore. Ergo no need
for further maintenance of striplibsfrommexfile.m, but we leave the
file there for documentation, should a similar need arise in the
future.

Unfortunately the same is not true for the MS-Windows or macOS
platform.