Add explicit dllexport spec to mex file build for Octave on Windows.
Suddenly, since Octave 8 for MS-Windows, we need to prefix our mexFunction()
with an explicit  __declspec(dllexport), or loading PsychHID.mex will not work anymore,
due to PsychHID.mex not exporting the mexFunction entry point! Why now, and why
only for PsychHID.mex, I don't know, but this fixes it. We must not do this on other
systems, or for Matlab, as Matlab's mex build script already defines linkage.

While we are at it, clean up a bit, and substitute EXP -> PTB_EXPORT for better clarity.