PsychOpenXR: Update code comments and diagnostic messages for HTC Vive Pro Eye.
Testing with the latest SteamVR 1.25.6 as of mid-March 2023 on Windows-10 with a
HTC Vive Pro Eye, with both AMD and NVidia gpu's, shows that the hang/crash bug
of SteamVR on MS-Windows with multi-threading and Screen('Begin/EndOpenGL')
is not only happening with the OculusVR backend and OculusVR compositor, but also
for natively SteamVR supported VR HMD's like this Vive, where the SteamVR compositor
is used for display, instead of an external 3rd party compositor.

This seems to be clearly a SteamVR-on-Windows immanent bug. Update diagnostic help
messages to also warn about the problem / give troubleshooting tips on SteamVR+Windows
with all backends.

Update other code comments with testing results of SteamVR Windows + HTC Vive Pro Eye.