PsychOpenXRCore: Change controller input mappings for HTC Vive controllers.
Testing with SteamVR on Windows with two HTC Vive controllers showed the not so great
current button mapping. Both menu buttons were mapped to the same Enter button.

Lets change this for compatibility with Oculus touch controllers and map the right controllers
menu button to the Home button instead.

Also expose the left/right trigger buttons - for pulling the triggers into their end-position,
as equivalent of the Oculus A and X buttons, so we can make use of this trigger click info.

We expose the two left/right system click buttons as B and Y buttons. However, at least
with SteamVR on Windows, these two buttons are not reported to us, seem to be system
private.

Other testing results show that the Vive Pro's volume up/down buttons work and get
reported to us, but the microphone mute button and system click button are not reported.

Also some minor dead code removal. Haptics pulse testing showed that the Vive controllers
do obey provided target pulse durations and amplitude under SteamVR + Windows. They
also provide a wider range of vibration frequencies. Seems to work at least in the tested
range of 100 Hz to about a maximum of 3300 or 3400 Hz, haven't tested lower than 100 Hz
frequency. This is a much larger/finer range than the two discrete values on the Oculus
Rift CV-1 touch controllers of 160 Hz and 320 Hz, or of 160 Hz and 500 Hz for the Rift-S
touch controllers.

This concludes baseline enablement, tweaking and testing for SteamVR on Windows with
the HTC Vive Pro Eye HMD and gives us a reasonably well working equivalent of what we
have on the Oculus Rift CV-1.