PsychOpenXR[Core]: Add basic gaze tracking and reporting support.
"[state, touch, gaze] = PsychOpenXRCore('GetTrackingState', ..., reqs)" now returns gaze
tracking data in the new 'gaze' struct array if reqs flag 4 is set.

The functions help is worded to allow for returning an array with one or multiple structs,
allowing for reporting of a single eye gaze vector, e.g., a synthesized cyclops eye gaze,
or two vectors for each of the two eyes, or even three for left, right, and cyclops eye,
depending on what underlying extensions are supported. Right now, only support for the
XR_EXT_eye_gaze_interaction extension is implemented, providing one gaze vector, which
is either synthesized from the two vectors of a binocular eye-tracker, or may be only one
vector reported for a monocular tracked eye.

This is so far successfully tested with a HTC Vive Pro Eye with HTC's SRAnipal runtime, which
utilizes a built-in Tobii binocular eye tracker of the Vive Pro Eye. The extension seems to be
implemented in a OpenXR api layer, as the same extension is also reported when running
under MonadoXR for Windows - iow. it is part of that 3rd party HTC layer, not of the core
XR runtime. This gives some hope that one could also use the extension with different XR
runtimes, but the eye tracking does nothing - does not deliver any samples - under Monado.
Maybe there is some implicit dependency on SteamVR for Windows, or maybe it is because
my Monado test dll does not actually access the Vive, so maybe it doesn't get powered up
sufficiently?

A future extension would be XR_FB_eye_social_interaction for binocular gaze tracking, but
I can't test this yet, as this extension is not supported on the given test hw+sw.

Limitations:

- No multi-threaded background gaze recording.
- Only one gaze vector, not per-eye.
- Only gaze reporting, no eye opening or pupil diameter or confidence value.
- No timestamps of actual gaze sample acquisition, but only interpolated timestamps.
- No callback driven sampling, just query of potentially interpolated results, based on
  real samples produced at unknown time-points by the hardware.

XR_FB_eye_social_interaction would provide binocular gaze vectors and some confidence
value, but likely low-pass filtered/smoothed samples as the purpose is for driving eye
animations for avatars, not for precise gaze data acquisition.

Currently as of end of March 2023, OpenXR does not have a research grade eye gaze
extension yet which would check all our boxes of wants and needs.

The PsychOpenXR('PrepareRender', ...) function currently exposes the tracking info as a
3rd optional return argument, a struct array of gaze samples, 1-3 elements for monocular/
cyclops eye, binocular, or both. Current code post-processes to provide the raw gaze vectors
from PsychOpenXRCore, a 7-component vector with eye-ball center [x,y,z], followed by a
quaternion encoding eyeball pose, together with a timestamp and status flag about tracking
status. Post-processed output is 4x4 transformation matrices HMD local and global for OpenGL
rendering or other 3D transformations, and a eye ray spec with position and direction vector,
 e.g., for 3D ray + object intersections. More to come...

The implementation and API of this is for now considered experimental and unstable, with
potentially arbitrary backwards incompatible changes pending.

This also allows for initial 3D gaze-ray to 2D gaze position mapping in window coords, ie. mapped from
eye gaze pose vectors to quadview or projectionview space, then to onscreen window drawbuffer
space.