VR demos: Remove obsolete/unpopular options, cleanup.
Remove task requirements keywords that are only applicable to the old PsychOculusVR
driver for Oculus Rift DK-1 and DK-2 under Linux/X11 and Windows with the old Oculus
VR 0.5-beta runtime. I assume that by now this is the least often used driver + hw combo,
so remove keywords from demos which are only useful with that driver, to declutter things
a bit. Some other cleanup.