[Legacy]XOrgConfCreator: Handle HMDs on NVidia proprietary different.
Default to "don't expose to X11" with LegacyXOrgConfCreator, add a user
choice for XOrgConfCreator.

The option needs to be "yes" for the NVidia proprietary driver with VR
HMDs if one wants to use those HMDs with our PsychOculusVR driver (Oculus
Rift DK1/DK2) or with our PsychOpenHMDVR driver (others). It must be "no"
for use with our OpenXR driver, or if one wants to use Vulkan direct
display mode at all on NVidia's proprietary driver, as otherwise RandR
output leasing will be disabled.

In other words: No is almost always the right choice, unless one wants
to use legacy VR drivers for some reason, e.g., because OculusVR is still
the best choice for Oculus Rift DK1 and DK2.