PsychOpenHMDVR: Add more hacks to deal with non-desktop flag.
Does not help a bit on the HTC Vive Pro Eye, or likely on any HTC Vive.
The specifics of how the Vive power-up sequence and non-desktop marking
and output detection interact make this essentially a no-go. But it does
supposedly still work with Oculus Rift, or so my past code and comments
claim.

All modern Linux kernels in all distributions we officially support do
mark VR HMD's as non-desktop, so they don't show up in Screen()'s output
enumeration. The hackery of manually overriding this did work for the
Oculus Rift DK-1/DK-2/CV-1, but the Vive seems to be a lost cause.

I think our OpenHMD driver needs a massive overhaul, or simply it needs
to go the way of all mortal things. It's just not really worth it to try
to patch this up again without anybody funding it.