PsychVRHMD: Skip OpenHMD in high-precision timing probe sequence.
Use of OpenHMD driver just got too fickly in setup and operation on
modern Linux - this is more of a liability than a feature. Don't use
it by default anymore iff a high timing precision driver is requested,
it just gets in the way. If somebody wants this fixed for a special use
case they can contract me to fix it for them.

The last drop in the bucket was trying to test this with the
HTC Vive Pro Eye, which is even more fickly than the Oculus devices...