DaqAInScan: Try to fix live data retrieval with more than 1 channel.
Cfe. https://psychtoolbox.discourse.group/t/sampling-frequency-of-analogue-input/4780/10

Can cause data loss or misalignment if DaqAInScanContinue(daq, options, 1);
for live data retrieval is used and more than one analog input channel is
in use: If number of new received HID reports is not an integral multiple
of the channel count, data would get truncated and then misalign the
channels.

We need to detect this case of insufficient new reports and then cut
off early enough, and save postfix reports for the next call to
DaqAInScan() for use a prefixed reports, so no data gets thrown
away needlessly.

This may or may not fix the problem - Untested.

-> The whole implementation of DaqAInScan is a mess. One can try
   to patch it up, but a rewrite from scratch would be the proper
   solution to retain some sanity.

Unpaid work time spent diagnosing and trying to fix: 4 hours!