PsychOpenHMDVR: Add HTC Vive (Pro) specific undistortion shader.
This adds a shader, ported from earlier versions of Monado for OpenGL 4
to a format usable with our code and OpenGL 2.1, which implements the
HTC Vive's DPOLY3 undistortion and color abberation correction warp.
Plus setup code that uses undistortion calibration data extracted from
the SteamVR Lighthouse config JSON file, specific to the rented HTC Vive
Pro Eye.

JSON config example location are at following locations:

On Linux:
.steam/debian-installation/config/lighthouse/lhr-b921a891/config.json

On Windows:
C:\Program Files (x86)\Steam\config\lighthouse\lhr-b921a891/config.json

The point of this was to improve image quality under OpenHMD for the
HTC Vive Pro Eye and other Vive HMDs. This was only partially successful.
Image quality is a bit improved, but not quite as good as on Monado or
SteamVR. And even with SteamVR the quality of undistortion in the
periphery is not that great.

All in all not a fantastic improvement, but the work has been done so
might as well merge it...