PsychOpenHMDVR: Improve X-Screen matching logic.
Not that it helps. Getting a dual-X-Screen setup with the HTC Vive Pro Eye
working is basically impossible. EDID override works to get a dual-X-Screen
setup going, but then video mode setting fails, or the desktop UI crashes,
or a hundred other random things. This is so painfully shaky that it is
practically unworkable.

I guess the good days of the OpenHMD driver are just in the past and we
should fully focus on Monado and OpenXR instead. Maybe there's a tiny bit
of use on Windows if timing/timestamping is needed for narrowly scoped
applications...