CalibrateMonSpd() et al.: Allow different colormeters, fix it on standard displays.
The old code only allows use of a PhotoResearch PR-650 colormeter, not of
any of the other colormeters supported by CMCheckInit() and friends! Fix
this. Also default meterType to zero, for simulated colormeter, so I can
test the code without access to such a device.

The old code only worked for CRS and VPixx visual stimulators, and was
utterly broken on standard gpu + display combos or other high-precision
displays than CRS / VPixx. We fix this by introducing a simulated clut
that has the same size of 256 slots as the real hardware clut's of the
various CRS and VPixx devices, so all the clut trickery in this code works
as expected. For that we can use the imaging pipelines PsychImaging()
task PsychImaging('AddTask', 'AllViews', 'EnableCLUTMapping', 256, 1);
to switch on the imaging pipeline on standard displays, and emulate a clut
compatible with the ones in above mentioned stimulators. The clut uses
256 slots like those, although we only need 2 slots in total, but this
makes the code easier to understand. We use high precision mode, so the
emulated clut has 23 bits effective linear output precision - plenty for
all existing current display hardware in the world.
This allows to remove various broken special casing and to get away from
the problem that the whole code assumes a 256 slot lut on 8 bpc framebuffers,
but modern gpu + os combos have anything from 256 slots to 512, 1024, 2048
or even 4096 slots, with framebuffers from 8 bpc to 10 bpc, 11 bpc, or
even 16 bpc.

Also fix some bugs with cal.describe.dacsize, which was fed totally wrong
values for the last 5 years at least. A proper dacsize is one that results
in a size of the gamma correction table compatible with the size of the
hardware clut used for gamma correction, ie. 256 for CRS and VPixx, and
whatever Screen('ReadnormalizedGammatable') reports on standard displays
for the gpu that drives the display.