Eyelink mex files for 64-Bit Octave/Matlab on Linux rebuilt against SDK v2.1.
-> Contains Eyelink('ImageTransfer') improvements.

Rebuilt mex files against the more recent eyelink_core SDK version 2.1+,
instead of the old 1.9 SDK.

This was built on a Ubuntu 20.04.6-LTS build system, and tested to work
on at least Ubuntu 20.04 and 22.04 LTS.

It requires libeyelink_core-2.0.so.0 from the current SDK/Runtime, but
also works with the older runtime if one creates a symlink from
libeyelink_core-2.0.so.0 to the existing old libeyelink_core.so.0.
However, upgrading the runtime is better, as definitely some bugs were
fixed, e.g., wrt. Eyelink dummy mode.