EyelinkPictureCustomCalibration, EyeLinkPicture: Use new Eyelink('ImageTransfer').
Use imread() to read the image file and pass it as image matrix to
Eyelink('ImageTransfer'), instead of the old style of passing the
image file name. This allows to load other than .bmp images, which is
a good thing, because the demos tried to use .jpg images, so they were
broken.

Also some cleanup in functions, and remove special cases for Octave
which are no longer needed on at least Octave 5.2+ as shipping in Ubuntu
20.04-LTS and later - the oldest version of Ubuntu we support as of PTB
3.0.19.

-> Only tested in Eyelink dummy mode due to lack of hardware.