PsychOpenXR/PsychOpenHMDVR: Add workaround for SRAnipal left<->right eye center switching bug.
The api seems to report left eye center in the right eye struct and vice versa, causing switched
eye center positions between the eyes. Fix this, and also negate the x-axis position (sign switch)
of the fused cyclops eye.

This removes almost all horizontal bias of the left and right eye tracker streams, and also
removes the jitter and instability in the cyclops eye fused stream when looking at the center
of the HMD. This may be as close as we can get to proper tracking with SRAnipal on HTC
HMDs.

Note that the same jitter exists in the "cyclops eye" single gaze data reported by OpenXR
XR_EXT_eye_gaze_interaction for the HTC Vive Pro Eye. The reason is almost certainly the same
as when using SRAnipal directly, as HTC implements its OpenXR extension as a simple wrapper
around its SRAnipal implementation - specifically the cyclops eye reporting. They fall victim to
the same left<->right eye center switching bug, amplifying cyclops eye jitter, but we can not do
anything about this in case of XR_EXT_eye_gaze_interaction. Such is life...