New MultiTouchPinchDemo.m: Implements simple pinch gesture on touch screen.
Motivated by the line-intersection task from:
https://psychtoolbox.discourse.group/t/help-with-multitouch/4953

Tested on Linux/X11 Ubuntu 20.04.6-LTS, and on Windows 11, on a
MS Surface Pro 6 tablet with single touchscreen.

User didn't pay the over 1.5 hours of work, but add the demo here now
"as is", to not let the work go to waste totally, now that multiple
months have passed since the inquiry.