BitsPlusIdentityClutTest: Disable encoder test if Vulkan display is used.
Current design leads to the tests running before the Vulkan backend is
fully in charge, so we display the PTB welcome screen or pixeltrash during
the test, instead of the test stims, which leads to false-positive test
failure. Just don't offer the test option under Vulkan.

DatapixxGPUDitherpatternTest is an alternative way to test with Vulkan,
which actually works for this purpose due to full Vulkan init before
running the test.