Screen/OSX: Remove conserveVRAM flag kPsychDontCacheTextures.
Use of texture client storage under macOS did make some sense in
the early years, when gpu's had only 16 - 128 MB of VRAM, but it
is pointless with gpu's supported by currently supported macOS
versions, where VRAM is measured in GB, not MB. I'm sure nobody
used this conserveVRAM flag 2 since years. Also this code seems to
trigger bugs in current macOS OpenGL, leading to various failures
and crashes, so it is pointless anyway - and performance was never
any good.

Remove this flag and code. The flag 2 will be recycled in the next
commit for something that may be more useful in some fringe use
cases.