Screen/OSX: Try to deal with special cases wrt. macOS timing.
The new flag 2 == kPsychDontSwitchToOptimalVidMode for
conserveVRAM settings modifies the behavior of the macOS
Intel visual timing workaround in fullscreen mode: Setting the
flag prevents choice of the auto-detected optimal video mode
for best visual stimulus timing. Instead it causes a fallback that
chooses the best suited mode that has the same resolution and
refresh rate as the current active video mode. This is meant to
maybe help with special external display devices, e.g., Matrox
DualHead2Go display splitters. The current code misbehaves and
selects a wrong mode. It is not clear if this new flag will help in that
case, as the original reporter of the problem didn't care enough to
actually help in diagnosing and finding a perfectly targeted solution.
So this flag and method is just based on a hunch about what might
be the cause and one possible (crude!) solution.

In general one does not want to use this flag.
See following forum post for reference:

https://psychtoolbox.discourse.group/t/open-window-cant-set-to-a-specified-rect-size-on-macos-with-matrox-dualhead2go/5061

We will probably never learn if this actually works.