Screen: Add info if init of drawtext plugin takes unusually long.
We time the duration of drawtext plugin init and output an info message
to the user if the init takes longer than usual, with "usual" defined for
now as 5 seconds. Long init could mean that fontconfig caches had to be
rebuilt, and this happening frequently could hint at a fontconfig cache
problem. Point user to 'help DrawTextPlugin' for troubleshooting tips in
this case.

This is essentially never a problem on Linux, only very rarely on macOS,
but has been more frequent on MS-Windows. Not sure if this pointer will
help, as many users don't read, but probably does not harm either.

Suggested by GitHub user @mirh in issue #786.