PsychPortAudio: Allow skip of acting on result of Pa_IsFormatSupported().
If verbosity is lowered from default 4 to (still verbose) level 3, then
the results of Pa_IsFormatSupported() are no longer evaluated and acted
on. Iow. the early error exit is avoided, letting the 'Open' function
proceed to Pa_OpenStream(). Pa_OpenStream() will do its own error checking
and error exit if the device can't be opened/initialized as desired.

The idea here is to allow for a way to bypass Pa_IsFormatSupported(), in
case its implementation in a Portaudio backend is buggy or flawed in some
way and causes a "false positive" report of trouble / invalid settings.
The validation code of Pa_IsFormatSupported() differs from the one that
is executed during device open/init, so those differences could matter in
some edge cases with special sound hardware or unusual setups. This allows
a means of easier debugging of such cases. Specifically we had a bunch
of reports on the forum which aborted with paBadIo... error in validation.
Maybe this helps narrow down the source of trouble some day.

Also some rewording of output messages and some code reformatting.

Also add diagnostic output handling for paSampleFormatNotSupported.