PsychPortAudio/Linux: Add workaround for timestamp failure with PulseAudio.
The initial Portaudio PulseAudio backend implementation has a limitation.
If a half-duplex playback stream is started, the first n paCallback
invocations will report invalid audio timestamps. This is probably due to
audio buffer priming at start of playback of an audio stream, where the
callback is called in quick succession to prime the output buffer with
data to play, and no timestamps are available. Having invalid timestamps
would cause misscheduled start of sound and misreported latency and start
timestamps, so we can not properly start in this situation.

Add a workaround: If Pulseaudio is used as backend and we are in this
initial run of invalid timestamps, just "no-op" by outputting silence.
Only once at least one valid timestamp is provided, breaking the run of
n invalid timestamps, will the driver go into normal sound scheduling
mode for playback and capture. This is implemented inside paCallback().
The 'Start' function will turn the first 'Start' since device open, ie.
a real startup of the audio stream, into a waiting start under Pulseaudio,
so 'Start' does not return until playback is really running with valid
timestamps. It will also reset the timestamp failure counter to zero, as
after our little workaround, those initial bad batch of n timestamps did
not have any ill effect in practice, so we can start a session with a
clean slate and zero effective timestamp failures.

n is typically correlated with samplerate, e.g., with around n = 140
at 44100 Hz, 152 at 48000 Hz, at least on my tested machine.

Some oddity is that we don't get invalid timestamps in pure capture, or
full-duplex mode, or apparently if in half-duplex capture is started
before playback. Only if half-duplex playback is started before capture,
or playback only. I guess some differences in the priming strategies, or
maybe if the sound hardware on the Pulseaudio server side is already
running at start time of a playback stream, so timestamps are already
available during priming. Didn't investigate further.