linuxmakeitoctave3, datapixxmakemex: Enable exception handling explicitely for RaspberryPi.
Add -fexceptions compiler option for building mex files for Octave
on RaspberryPi OS. For some reason this flag is there implicitely
on standard Ubuntu Linux builds, but not on RaspberryPi OS. The
result was missing C++ exception handling for mex files build for
the RaspberryPi, so an error abort in a mex file always ended with
an unhandled C++ exception which terminated octave as a whole.

This fixes it after rebuilding all mex files for RPi.