PsychOpenXR: Make special eye tracking y-axis hack conditional.
Assume it is only needed on HMD's under Windows which use SRAnipal eye
gaze tracking, e.g., the HTC Vive Pro Eye for which we need the y-axis inversion
hack.

Only time - and testing on other devices - will tell if this hack is more generally
needed or not, iow. if it might not be a hack but a mathematical necessity.