Screen: Optimize storage format for LUMINANCE16 textures for texture normalization.
In PsychNormalizeTextureOrientation(), when converting non-renderable luminance
textures to rgba textures, and also converting planar luminance textures to packed
standard textures, we mapped LUMINANCE16[_ALPHA16] to RGBA32F, which is very
wasteful with video memory, blowing up texelsize from 2 or 4 Bytes to 16 Bytes.
Instead handle the LUMINANCE16[_ALPHA16] cases and map them to RGBA16,
for cut down to 8 Bytes, ie. cutting texture memory consumption to half.

Doesn't change the fact that converting luminance to upright orientation for
render-to-texture, or for simpler processing via Screen('TransformTexture'), or
converting to packed pixel format from planar is somewhat expensive in compute
and memory and memory bandwidth, and should be avoided unless neccessary,
but at least it is not extra wasteful for no reason.