Screen: Rewrite CRS clut update function RenderClutBits++ as a workaround.
Switch from using point rendering via glVertex2i() of the clut T-Lock update pattern
to instead using glRasterPos2i + glDrawPixels().

Why? Because the proprietary AMD OpenGL driver on Windows has bugs in
glVertex2i positioning when rendering to the OpenGL system backbuffer. The
most easy way to work around this atm. is to use glRasterPos2i + glDrawPixels()
to avoid the glVertex2i AMD bugs. This has the added benefit that we now use
this glRasterPos2i + glDrawPixels() combo consistently for all CRS T-Lock rendering,
ie. FE1 stereo driving and DIO codes, and for identity pixel passthrough tests and
gamma table tweaking in PsychGPUTestAndTweakGammaTables(). At least we only
have to think about one potentially broken function (glRasterPos2i()) in the future,
not two separate ones.

This was found on AMD RavenRidge Vega11 under Windows 10 with AMD driver
23.11.1. No such problem happened with the same hardware on Linux during my
testing, but there is issue #807 describing similar symptoms on AMD + Linux for
some other AMD gpu's. So the root cause is not quite clear, except that the buggy
functions seem to not be buggy when rendering into a FBO, e.g., when using the
Vulkan backend, but only be buggy when rendering into the OpenGL backbuffer.